+++ src/index.js
@@ -18,6 +18,7 @@ import { logout } from "./logout.js";
 import { setMode } from "./mode.js";
 // import { setIframe } from "./iframe.js";
 import { user, setUser } from "./user.js";
+import "./user.update.js";
 import { isTestHostname } from "./utils.js";
 import { refresh } from "./refresh.js";
